Place footer below main content (#4014)
Update the site's layout implementation so that the main content can
scroll with the entire viewport rather than having a separate scroll
view. This allows the footer to be placed after the main content rather
than fixed at the bottom and better aligns the site's layout with
dart.dev and docs.flutter.dev.

- Don't limit the body to a height of 100%, so that the main content can
be scrolled as part of the primary/full viewport. This is what allows
the footer to be after the content.
- Make header sticky so they always remain visible at the top of the
viewport.
- Make sidebars sticky, but below the header, so that they remain
visible when scrolling the main content.
- Remove custom webkit scrollbar styles as they cause part of the
header's chrome to go missing and are platform dependent.
- Add `color-scheme: <light or dark>` to the theme styles so the browser
provides properly themed scrollbars.
- Removing these styles also resolves a TODO relating to the mobile
version of the sidebars.

Resolves https://github.com/dart-lang/dartdoc/issues/4013