Simplify Inheritable.computeCanonicalEnclosingContainer. (#4047)
This method is very complicated, so any simplification is handy.

* It used to start with a very long if-then body, and the else body was
rather lost far below. Instead we can implement some simple short
circuiting, if we're not dealing with an inherited element.
* It used to carry around a `found` variable, breaking the loop when it
found a container. But it's simpler to just return at that point.
* The final return value is always `null`, representing a state where no
canonical enclosing container is found. Just return `null` to simplify
understandability.

Otherwise this is a no-op.