Simplify InheritingContainer wrt inherited elements (#4059)
While working with analyzer's InheritanceManager3, I noticed there was a
fair bit of code in dartdoc's InheritingContainer that could be
simplified. The `_inheritedElements` field is relied on by
`inheritedMethods`, `inheritedOperators`, and `allFields`. It performs
some calculations to come up with the collection of inherited elements
using its own homegrown dartdoc logic, but this logic exists in the
analyzer APIs on `InterfaceElement`. We can basically replace this field
with access to `element.inheritedElements`.

While investigating, I found that I could make a few members
`@visibleForTesting`: `inheritedMethods`, `inheritedOperators`. Also
`allFields` can be made private. And `Enum.hasPublicEnumValues` can be
removed, as it is the same as the super implementation. This leads to
all of the deleted code in `templates.runtime_renderers.dart`.