Refactor handling of generateDocs:false (#4095)
Not sure if anyone is using this option really. But I think the handling
can be simplified.

But it seems that having multiple Generator classes just for deciding
whether to write files or not is excessive, and likely to get out of
sync (does the EmptyGenerator really still do whatever GeneratorFrontend
does, just without writing files?)