Rename LanguageFeature and remove FeatureSet mixin (#4099)
`LanguageFeature` was introduced in
https://github.com/dart-lang/dartdoc/pull/2221/ for chips indicating
null safety. Since we now use chips for other purposes, such as class
modifiers, the name is no longer descriptive.

This change renames it to `Tag` to better model the general concept of a
highlightable tag on an element. This is somewhat related to`Attribute`
, and the two could potentially be merged in a future step.

Furthermore, `FeatureSet` is removed and its functionality is inlined
into `ModelElement`.

Finally `ContainerModifier` is extended to store the class modifier
description and URL directly, instead of in separate top level maps .