Bump updateAnalysisOptions parameter for deprecation (#4109)
When analyzer 8.2.0 was published, with a new deprecation, it started
this cascade of issues:

1. We currently run two SDKs in CI: one set of tests uses Dart 3.8.0,
and another uses Dart dev (3.10-something). Those dev bots pick up
analyzer 8.2.0, which triggers a CI failure due to the deprecated
`updateAnalysisOptions3` parameter.
2. In order to stop using that parameter and instead use
`updateAnalysisOptions4`, we have to always get analyzer 8.2.0, because
analyzer 8.1.1 does not have the new parameter.
3. In order to get analyzer 8.2.0, we need Dart 3.9.0. So this PR
changes our CI to use Dart 3.9.0 and Dart dev.